rename `config` to `plugin`
