+++ source-code/sdk-js/package.json
@@ -21,7 +21,7 @@
 		".": "./dist/core/index.js",
 		"./adapter-sveltekit": "./dist/adapter-sveltekit/index.js",
 		"./adapter-sveltekit/*": "./dist/adapter-sveltekit/utils/*/index.js",
-		"./config": "./dist/config/index.js",
+		"./plugin": "./dist/plugin/index.js",
 		"./detectors": "./dist/detectors/index.js",
 		"./detectors/*": "./dist/detectors/detectors/*/index.js",
 		"./runtime": "./dist/runtime/index.js"
